requirements.txt
torch>=2.0.0
diffusers>=0.30.0
transformers>=4.30.0
accelerate>=0.20.0
pillow>=9.0.0
numpy>=1.21.0
vtk>=9.0.0
matplotlib>=3.5.0
pandas>=2.0.0
tabulate>=0.9.0
